Remove unused files, split controllers into two files, created database access layer, finished poetry setup
