+++ .gitignore
@@ -65,4 +65,6 @@ target/
 #Ipython Notebook
 .ipynb_checkpoints
 
-podaac/controllers/__pycache__/
+hydrocronapi/controllers/__pycache__/
+
+.idea
