+++ README.md
@@ -9,10 +9,16 @@ as CSV and geoJSON.
 Python 3.10+
 
 ## Usage
+Before starting the server you must first start a local database instance. The easiest method is to use docker
+
+```
+docker run --name hydrocrondb -e MYSQL_DATABASE=test -e MYSQL_ROOT_HOST='%' -e MYSQL_ROOT_PASSWORD=my-secret-pw -p 3306:3306 -v $(pwd)/mysql/20230601_test.sql:/docker-entrypoint-initdb.d/20230601_test.sql -d --rm mysql:latest
+```
+
 To run the server, please execute the following from the root directory:
 
 ```
-python3 -m podaac
+python3 -m hydrocronapi
 ```
 
 and open your browser to here:
@@ -33,8 +39,8 @@ To run the server on a Docker container, please execute the following from the r
 
 ```bash
 # building the image
-docker build -t podaac .
+docker build -t hydrocronapi .
 
 # starting up a container
-docker run -p 8080:8080 podaac
+docker run -p 8080:8080 hydrocronapi
 ```
