--- docker/docker-entrypoint.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-set -e
-
-
