+++ hydrocronapi/swagger/swagger.yaml
@@ -108,7 +108,7 @@ paths:
                 type: array
                 items:
                   type: string
-      x-openapi-router-controller: podaac.controllers.default_controller
+      x-openapi-router-controller: hydrocronapi.controllers.timeseries
   /timeseriesSubset:
     get:
       summary: Subset by time series for a given spatial region
@@ -208,18 +208,5 @@ paths:
                 type: array
                 items:
                   type: string
-      x-openapi-router-controller: podaac.controllers.default_controller
-components:
-  schemas: {}
-  securitySchemes:
-    application:
-      type: oauth2
-      flows:
-        clientCredentials:
-          tokenUrl: http://example.com/oauth/token
-          scopes:
-            write: allows modifying resources
-            read: allows reading resources
-      x-tokenInfoFunc: podaac.controllers.authorization_controller.check_application
-      x-scopeValidateFunc: podaac.controllers.authorization_controller.validate_scope_application
+      x-openapi-router-controller: hydrocronapi.controllers.subset
