--- podaac/models/__init__.py
@@ -1,5 +1,0 @@
-# coding: utf-8
-
-# flake8: noqa
-from __future__ import absolute_import
-# import models into model package
