+++ tests/__init__.py
@@ -3,14 +3,11 @@ import logging
 import connexion
 from flask_testing import TestCase
 
-from podaac.encoder import JSONEncoder
-
 
 class BaseTestCase(TestCase):
 
     def create_app(self):
         logging.getLogger('connexion.operation').setLevel('ERROR')
-        app = connexion.App(__name__, specification_dir='../swagger/')
-        app.app.json_encoder = JSONEncoder
+        app = connexion.App(__name__, specification_dir='../hydrocronapi/swagger/')
         app.add_api('swagger.yaml')
         return app.app
